[REFACTOR] 파일명 변경 및 불필요한 파일 삭제
