+++ src/apis/auction/auctionDeckApi.ts
@@ -1,5 +1,5 @@
 import instance from "@/apis/instance";
-import type { DeckAuctionResponse } from "@/types/auction/deck";
+import type { DeckAuctionResponse } from "@/types/auction/deckApi.type";
 
 export const getAuctionDeck = async (): Promise<DeckAuctionResponse> => {
   const res = await instance.get<DeckAuctionResponse>("/auctions/deck");
