+++ src/apis/my/myAuctionApi.ts
@@ -1,5 +1,5 @@
 import instance from "@/apis/instance";
-import { MyAuctionsResponse } from "@/types/auction/myauction";
+import { MyAuctionsResponse } from "@/types/auction/myauctionApi.type";
 
 export const getMyAuctions = async (): Promise<MyAuctionsResponse> => {
   const res = await instance.get<MyAuctionsResponse>("/auctions/my");
