+++ src/hooks/auction/useCreateAuctionApi.ts
@@ -1,6 +1,6 @@
 // src/hooks/auction/useCreateAuction.ts
 import { createAuction } from "@/apis/my/createAuctionApi";
-import { CreateAuctionRequest } from "@/types/auction/auction";
+import { CreateAuctionRequest } from "@/types/auction/auctionApi.type";
 import { useState } from "react";
 
 export const useCreateAuction = () => {
