+++ src/pages/home/HomePage.tsx
@@ -1,7 +1,7 @@
 import LoadingSpinner from "@/components/common/LoadingSpinner";
 import HeaderBell from "@/components/mainpage/HeaderBell";
 import SwipeDeck from "@/components/mainpage/SwipeDeck";
-import { useAuctionDeck } from "@/hooks/auction/useAuctionDeck";
+import { useAuctionDeck } from "@/hooks/auction/useAuctionDeckApi";
 
 export default function HomePage() {
   /** 실제 경매 카드를 가져오는 API 훅 */
