+++ src/pages/search/SearchResultPage.tsx
@@ -1,7 +1,7 @@
 // SearchResultPage.tsx
 import { Delete, Goback, Search } from "@/assets/svgs/search";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
-import { useSearchAll } from "@/hooks/search/useSearch";
+import { useSearchAll } from "@/hooks/search/useSearchApi";
 import React, { useEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
