+++ src/types/auction/auctionApi.type.ts
