+++ src/types/auction/bidApi.type.ts
