+++ src/types/auction/deckApi.type.ts
