--- src/types/auction/image.ts
@@ -1,9 +1,0 @@
-export interface ImageUploadResponse {
-  success: boolean;
-  status: number;
-  message: string;
-  timestamp: string;
-  result: {
-    url?: string;
-  };
-}
