--- src/types/auction/swipe.ts
@@ -1,10 +1,0 @@
-export type SwipeActionType = "DISLIKE" | "HOLD" | "BIDDING";
-
-export interface SwipeActionRequest {
-  auctionId: number;
-  action: SwipeActionType;
-}
-
-export interface SwipeActionResponse {
-  success: boolean;
-}
