+++ src/types/item/detail/itemDetail.type.ts
@@ -1,0 +1,1 @@
+export type ViewState = "LIVE" | "LIVE_BIDDING" | "WON" | "LOST" | "ENDED";
