--- src/types/my/pointHistory.type.ts
@@ -1,7 +1,0 @@
-export interface PointItemData {
-  id: number;
-  type: "CHARGE" | "USE";
-  title?: string;
-  date: string;
-  amount: number;
-}
